Refactor Region Data content into api/data
