+++ common/src/main/java/terrablender/api/data/BiomeMapping.java
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package terrablender.api;
+package terrablender.api.data;
 
 import com.google.common.collect.ImmutableList;
 import com.mojang.serialization.Codec;
