+++ common/src/main/java/terrablender/core/Registries.java
@@ -20,7 +20,7 @@ package terrablender.core;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.Identifier;
 import net.minecraft.resources.ResourceKey;
-import terrablender.api.RegionDefinition;
+import terrablender.api.data.RegionDefinition;
 
 public class Registries {
     public static final ResourceKey<Registry<RegionDefinition>> REGION = ResourceKey.createRegistryKey(Identifier.fromNamespaceAndPath(TerraBlender.MOD_ID, "region"));
